# MW-1028: Improved performance of survey logic file