Huge survey
survey logic file 3.5m - 21.86s
survey logic printable 3.4m - 28s
Big Survey
survey logic file 12.63s - 4.99s
survey logic printable 13s - 6s