@Shnoulle 2024-02-19
The survey logic file is THE place where we need an updated expression manager.

I think it's already done by `$_SESSION['LEMforceRefresh']` actually. But is it a good solution to disable evaluation when previewmode are logic, same with database? And the best solution seems to be this one:

- All place where LimeExpressionManager->sPreviewMode = 'logic', replace by a LimeExpressionManager::SetDirtyFlag()
- All place where LimeExpressionManager->sPreviewMode = 'database', replace by a LimeExpressionManager::SetDirtyFlag()

We can do the same for question and questiongroup and deprecate sPreviewMode after. using only LimeExpressionManager::SetDirtyFlag() (or another function)