@AmeniChakroun 2024-02-20
Please How to run the code (command i need)?