@lajosarpad 2024-02-20
> The survey logic file is THE place where we need an updated expression manager.
> 
> I think it's already done by `$_SESSION['LEMforceRefresh']` actually. But is it a good solution to disable evaluation when previewmode are logic, same with database? And the best solution seems to be this one:
> 
> * All place where LimeExpressionManager->sPreviewMode = 'logic', replace by a LimeExpressionManager::SetDirtyFlag()
> * All place where LimeExpressionManager->sPreviewMode = 'database', replace by a LimeExpressionManager::SetDirtyFlag()
> 
> We can do the same for question and questiongroup and deprecate sPreviewMode after. using only LimeExpressionManager::SetDirtyFlag() (or another function)

The idea is good. Currently we need to fix the performance issues so that Limesurvey will be quicker for the customers. We will try to return to code quality issues as well, as per the suggestion. We will need to be careful while doing so, because parts of the code rely on these preview values, such as

```
            if ($this->sPreviewMode == 'logic') {
                $this->subQrelInfo[$questionNum][$rowdivid]['aWarnings'] = $aWarnings;
            }
```

so in order to implement this improvement, we will also need to do some refactoring and further tests. I initiated discussions about this in the team.