@lajosarpad 2024-02-20
> Please How to run the code (command i need)?

@AmeniChakroun in order to try this improvement, you can

1. Visit <yourroot>/index.php?r=surveyAdministration/listsurveys  (the survey list page)
2. Click on a survey
3. Click on Tools
4. Click on Survey logic file
5. Wait for the page to load and measure its loading length
![image](https://github.com/LimeSurvey/LimeSurvey/assets/2805488/5c65d025-e20f-4b8b-8721-facc2089704d)

The second test is to simply click on the Open printable view button on the survey logic file page: 
![image](https://github.com/LimeSurvey/LimeSurvey/assets/2805488/d3293d0a-0ce5-4732-b3d1-414689fd15b4)