@lajosarpad 2024-02-20
> the code is paid or free (license) ?

This code is free, you can find the branch of this fix at https://github.com/LimeSurvey/LimeSurvey/tree/bug/MW-1028-survey-logic-file-performance even if you are not logged in to GitHub and you can clone it. You can test this by running

```
git pull
git checkout bug/MW-1028-survey-logic-file-performance
```

in your local repository and then

```
git checkout master
```

if you want to test the current performance in the way I described in my previous comment and, you can repeat the same test with the better performing new code if you

```
git checkout bug/MW-1028-survey-logic-file-performance
```